requirements.txt
accelerate
addict
altair
annotated-types
ansi2html
antlr4-python3-runtime
appdirs
astor
bids
bids-validator
blinker
blis
braceexpand
Braindecode
brotlipy
cachetools
catalogue
cfgv
click
cloudpathlib
confection
ConfigArgParse
contexttimer
contourpy
cycler
cymem
dash
dash-core-components
dash-html-components
dash-table
decorator
diffusers
distlib
docker-pycreds
docopt
docstring-inheritance
easydict
einops
fairscale
filelock
Flask
fonttools
formulaic
fqdn
fsspec
ftfy
gitdb
GitPython
greenlet
h11
h5py
huggingface-hub
identify
imageio
imageio-ffmpeg
importlib-metadata
interface-meta
iopath
ipywidgets
isoduration
itsdangerous
jedi
joblib
json5
jsonpointer
jsonschema
jupyter
jupyter-console
jupyter-client
jupyter-core
jupyter-server
jupyter-server-terminals
jupyterlab
jupyterlab-pygments
jupyterlab-server
jupyterlab-widgets
kaggle
kiwisolver
langcodes
lazy-loader
llvmlite
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
mne
mne-bids
moviepy
murmurhash
nbclient
nbconvert
nbformat
nest-asyncio
clip-retrieval
dalle2-pytorch
open_clip_torch
info-nce-pytorch
pytorch-msssim
reformer_pytorch
wandb
